Implement recursive definition support
Enables definition of a recursive function through defining a base case
using an `if` statement. Previously, doing this would produce compilable
code, but running the product executable would cause a stack overflow,
since the executable would attempt to evaluate both sides of the branch.

Making the `if` statement a built-in part of the language is a bit
crude, but is simple and works. Maybe, if time allows, this will be
revisited.