Enable -Weverything
Clang has a -Weverything option, that enables all compile-time warnings
and elevates them all to errors. This option is now passed to the
compiler, and all warnings/errors the code produced have been addressed.

Using -Weverything locks the project into using clang as its compiler:
this should not be a problem since clang is probably being used to
compile the LLVM output of the project anyway. If support for other
compilers is required, it should be simple enough to replace
-Weverything with an equivalent collection of portable flags.

Explicit Makefile rules for compiling the C files produced by flex/bison
are now included in the project since these C files absolutely do not
pass -Weverything checks.