Added int width specifiers
The width of an int can be specified at the site of its creation. For
example, the following function returns an int of width 2 bits and value
1:

[def two_bit_int []
    (2 1)]

Ints can still be written the old way; in these cases the compiler looks
for a default int bit-width (specified with the -w flag). If a default
has not been defined, the compiler throws an error.