Fix issue with nested case statements
Previously, nested case statements would cause phis to be given
incorrect incoming blocks, causing issues when compiling the LLVMIR
with clang.

To fix this, all case branches are given a second "exit" block
that they jump to before jumping to the phi block. This way, we
always know the block that was just left to get to the phi.